+++ source/ed.vcxproj.filters
@@ -145,6 +145,9 @@
     <ClInclude Include="ed\kit\gateway_impl.h">
       <Filter>kit</Filter>
     </ClInclude>
+    <ClInclude Include="ed\kit\module_impl.h">
+      <Filter>kit</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="ed\exceptions\exception.cpp">
@@ -211,6 +214,12 @@
       <Filter>kit</Filter>
     </ClCompile>
     <ClCompile Include="ed\kit\gateway_impl.cpp">
+      <Filter>kit</Filter>
+    </ClCompile>
+    <ClCompile Include="ed\kit\module_impl_callback.cpp">
+      <Filter>kit</Filter>
+    </ClCompile>
+    <ClCompile Include="ed\kit\module_impl.cpp">
       <Filter>kit</Filter>
     </ClCompile>
   </ItemGroup>
