+++ source/ed/kit/event_result.h
@@ -18,6 +18,7 @@ namespace ed
     EVENT_RING r;
 
     friend class module;
+    friend class module_impl;
     event_result( message &, module &, int local_id, bool result, EVENT_RING notify );
   public:
     event_result( const event_result & );
